refactor: encapsulate cluster generation tasks
